+++ src/kernel/fs/dev/file.zig
@@ -15,7 +15,7 @@ pub const DevFile = struct {
         return to_write;
     }
 
-    fn read(base: *fs.File, buf: [*]u8, size: u32) !u32 {
+    fn read(base: *fs.File, buf: [*]u8, size: usize) !usize {
         const ino = base.inode.getImpl(DevInode, "base");
         var to_read = size;
         if (to_read > ino.buff.len - base.pos)
